requirements.txt
flask
torch
diffusers
transformers
accelerate
Pillow
uuid

# pip install -r requirements.txt